Revert "Move `Service*` aliases to check engine"
This reverts commit e87d383f6a59b4b93b96b122acb9b6e415a70d05.
