+++ cmk/automations/results.py
@@ -19,7 +19,7 @@ from cmk.utils.labels import HostLabel, HostLabelValueDict, Labels
 from cmk.utils.notify_types import NotifyAnalysisInfo, NotifyBulks
 from cmk.utils.plugin_registry import Registry
 from cmk.utils.rulesets.ruleset_matcher import LabelSources, RulesetName
-from cmk.utils.type_defs import CheckPluginNameStr, Item, ServiceName
+from cmk.utils.type_defs import CheckPluginNameStr, Item, ServiceDetails, ServiceName, ServiceState
 
 from cmk.snmplib.type_defs import AgentRawData  # pylint: disable=cmk-module-layer-violation
 
@@ -27,7 +27,6 @@ from cmk.checkengine.checkresults import MetricTuple
 from cmk.checkengine.discovery import DiscoveryResult as SingleHostDiscoveryResult
 from cmk.checkengine.legacy import LegacyCheckParameters
 from cmk.checkengine.parameters import TimespecificParameters
-from cmk.checkengine.submitters import ServiceDetails, ServiceState
 
 DiscoveredHostLabelsDict = dict[str, HostLabelValueDict]
 Gateway: TypeAlias = tuple[
