+++ cmk/bi/lib.py
@@ -36,9 +36,7 @@ from cmk.utils.hostaddress import HostName
 from cmk.utils.macros import MacroMapping, replace_macros_in_str
 from cmk.utils.rulesets.ruleset_matcher import TagCondition
 from cmk.utils.tags import TagGroupID, TagID
-from cmk.utils.type_defs import ServiceName
-
-from cmk.checkengine.submitters import ServiceDetails, ServiceState
+from cmk.utils.type_defs import ServiceDetails, ServiceName, ServiceState
 
 from cmk.bi.type_defs import (
     ActionConfig,
